色の調整、コメント修正、cast追加

+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/CCharacter/CPlayer/PlayerState/PlayerActionState/PlayerFallingState/CPlayerFallingState.cpp
@@ -35,7 +35,7 @@ CPlayerFallingState::~CPlayerFallingState()
 void CPlayerFallingState::Enter(CPlayer& pPlayer)
 {
 	//攻撃の開始時間を取得.
-	m_StartTime = CTimeManager::GetTotalTime();
+	m_StartTime = static_cast<float>(CTimeManager::GetTotalTime());
 
 	//傾き角度の初期化.
 	m_CurrentTiltAngle = 0.f;
@@ -71,7 +71,7 @@ void CPlayerFallingState::Update(CPlayer& pPlayer)
 	pPlayer.SetTurnState(std::make_unique<CPlayerTurnIdleState>());
 
 	//経過時間を取得.
-	float t = CTimeManager::GetTotalTime();
+	float t = static_cast<float>(CTimeManager::GetTotalTime());
 
 	//プレイヤーの位置を取得.
 	D3DXVECTOR3 playerPos = pPlayer.GetPosition();
@@ -134,7 +134,7 @@ void CPlayerFallingState::Update(CPlayer& pPlayer)
 	pPlayer.GetPlayerLeftHand().SetPosition(leftHandPos);
 
 	//1フレームの速さを取得.
-	float dt = CTimeManager::GetDeltaTime();
+	float dt = static_cast<float>(CTimeManager::GetDeltaTime());
 
 	//攻撃された情報の取得.
 	m_Velocity.y += -m_Gravity * dt;
