+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/CCharacter/CPlayer/PlayerState/PlayerActionState/PlayerPushedState/CPlayerPushedState.cpp
@@ -41,7 +41,7 @@ void CPlayerPushedState::Enter(CPlayer& pPlayer)
 	m_StartQuat = pPlayer.GetQuaternion();
 
 	//攻撃の開始時間を取得.
-	m_StartTime = CTimeManager::GetTotalTime();
+	m_StartTime = static_cast<float>(CTimeManager::GetTotalTime());
 
 	//手の開始位置の設定.
 	m_RightHandStartPos = pPlayer.GetPlayerRightHand().GetOffsetPos();
@@ -66,7 +66,7 @@ void CPlayerPushedState::Update(CPlayer& pPlayer)
 	pPlayer.SetTurnState(std::make_unique<CPlayerTurnIdleState>());
 
 	//ゲーム全体の経過時間.
-	float t = CTimeManager::GetTotalTime();
+	float t = static_cast<float>(CTimeManager::GetTotalTime());
 
 	//現在の経過時間と開始時間の差が終了時間を上回ったら.
 	if (t - m_StartTime > m_EndTime)
